+++ src/SLOPE.cpp
@@ -56,23 +56,25 @@ List cppSLOPE(T& x, mat& y, const List control)
   standardize(x, x_center, x_scale, intercept, center, scale);
 
   auto lambda = as<vec>(control["lambda"]);
-  auto sigma  = as<vec>(control["sigma"]);
+  auto alpha  = as<vec>(control["alpha"]);
   auto lambda_type = as<std::string>(control["lambda_type"]);
-  auto sigma_type = as<std::string>(control["sigma_type"]);
-  auto lambda_min_ratio = as<double>(control["lambda_min_ratio"]);
+  auto alpha_type = as<std::string>(control["alpha_type"]);
+  auto alpha_min_ratio = as<double>(control["alpha_min_ratio"]);
   auto q = as<double>(control["q"]);
-  const uword n_sigma = sigma.n_elem;
-  double sigma_max = 0;
+  const uword path_length = alpha.n_elem;
+  double alpha_max = 0;
 
-  regularizationPath(sigma,
+  regularizationPath(alpha,
                      lambda,
-                     sigma_max,
+                     alpha_max,
                      x,
                      y,
+                     x_scale,
                      y_scale,
                      lambda_type,
-                     sigma_type,
-                     lambda_min_ratio,
+                     alpha_type,
+                     scale,
+                     alpha_min_ratio,
                      q,
                      family_choice,
                      intercept);
@@ -87,22 +89,22 @@ List cppSLOPE(T& x, mat& y, const List control)
                             tol_rel,
                             verbosity);
 
-  cube betas(p, m, n_sigma, fill::zeros);
+  cube betas(p, m, path_length, fill::zeros);
   mat beta(p, m, fill::zeros);
 
   uword n_variables = 0;
-  uvec n_unique(n_sigma);
+  uvec n_unique(path_length);
 
   mat linear_predictor = x*beta;
 
   double null_deviance = 2*family->primal(y, linear_predictor);
-  vec deviance_ratios(n_sigma);
-  vec deviances(n_sigma);
+  vec deviance_ratios(path_length);
+  vec deviances(path_length);
   double deviance_change{0};
 
   mat beta_prev(p, m, fill::zeros);
 
-  uvec passes(n_sigma);
+  uvec passes(path_length);
   std::vector<std::vector<double>> primals;
   std::vector<std::vector<double>> duals;
   std::vector<std::vector<double>> timings;
@@ -114,7 +116,7 @@ List cppSLOPE(T& x, mat& y, const List control)
   mat pseudo_gradient_prev(n, m);
 
   // sets of active predictors
-  field<uvec> active_sets(n_sigma);
+  field<uvec> active_sets(path_length);
   uvec active_set = regspace<uvec>(0, p-1);
   uvec strong_set;
   uvec previous_set;
@@ -144,7 +146,7 @@ List cppSLOPE(T& x, mat& y, const List control)
 
   uword k = 0;
 
-  while (k < n_sigma) {
+  while (k < path_length) {
 
     violations.clear();
 
@@ -154,11 +156,11 @@ List cppSLOPE(T& x, mat& y, const List control)
 
       gradient_prev = family->gradient(x, y, x*beta_prev);
 
-      double sigma_prev = (k == 0) ? sigma_max : sigma(k-1);
+      double alpha_prev = (k == 0) ? alpha_max : alpha(k-1);
 
       strong_set = strongSet(gradient_prev,
-                             lambda*sigma(k),
-                             lambda*sigma_prev,
+                             lambda*alpha(k),
+                             lambda*alpha_prev,
                              intercept);
 
       previous_set = find(any(beta_prev != 0, 1));
@@ -195,17 +197,19 @@ List cppSLOPE(T& x, mat& y, const List control)
 
         vec eigval = eig_sym(xx);
 
-        if (lambda.max()*sigma(k) == 0) {
+        if (lambda.max()*alpha(k) == 0) {
           rho = eigval.max();
         } else {
           rho =
-            std::pow(eigval.max(), 1/3)*std::pow(lambda.max()*sigma(k), 2/3);
+            std::pow(eigval.max(), 1/3)*std::pow(lambda.max()*alpha(k), 2/3);
         }
 
-        if (n < p)
+        if (n >= p) {
+          xx.diag() += rho;
+        } else {
           xx /= rho;
-
-        xx.diag() += rho;
+          xx.diag() += 1;
+        }
 
         U = chol(xx);
         L = U.t();
@@ -214,7 +218,7 @@ List cppSLOPE(T& x, mat& y, const List control)
       }
 
       res =
-        family->fit(x, y, beta, z, u, L, U, xTy, lambda*sigma(k), rho, solver);
+        family->fit(x, y, beta, z, u, L, U, xTy, lambda*alpha(k), rho, solver);
       passes(k) = res.passes;
       beta = res.beta;
 
@@ -251,11 +255,11 @@ List cppSLOPE(T& x, mat& y, const List control)
 
             vec eigval = eig_sym(xx);
 
-            if (lambda.max()*sigma(k) == 0) {
+            if (lambda.max()*alpha(k) == 0) {
               rho = eigval.max();
             } else {
               rho = std::pow(eigval.max(), 2/3)*
-                std::pow(lambda.max()*sigma(k), 1/3);
+                std::pow(lambda.max()*alpha(k), 1/3);
             }
 
             if (x_subset.n_rows < x_subset.n_cols)
@@ -283,7 +287,7 @@ List cppSLOPE(T& x, mat& y, const List control)
                             L,
                             U,
                             xTy,
-                            lambda.head(n_active)*sigma(k),
+                            lambda.head(n_active)*alpha(k),
                             rho,
                             solver);
 
@@ -308,7 +312,7 @@ List cppSLOPE(T& x, mat& y, const List control)
                                            x_subset*beta.rows(strong_set));
           uvec tmp = kktCheck(gradient_prev,
                               beta.rows(strong_set),
-                              lambda.head(n_strong)*sigma(k),
+                              lambda.head(n_strong)*alpha(k),
                               tol_infeas,
                               intercept);
           uvec strong_failures = strong_set(tmp);
@@ -325,7 +329,7 @@ List cppSLOPE(T& x, mat& y, const List control)
           gradient_prev = family->gradient(x, y, x*beta);
           uvec tmp = kktCheck(gradient_prev,
                               beta,
-                              lambda*sigma(k),
+                              lambda*alpha(k),
                               tol_infeas,
                               intercept);
 
@@ -386,17 +390,19 @@ List cppSLOPE(T& x, mat& y, const List control)
       }
     }
 
-    if (n_unique(k) > max_variables)
+    if (n_unique(k) > max_variables) {
+      k++;
       break;
-
-    k++;
+    }
 
     checkUserInterrupt();
+
+    k++;
   }
 
   betas.resize(p, m, k);
   passes.resize(k);
-  sigma.resize(k);
+  alpha.resize(k);
   n_unique.resize(k);
   deviances.resize(k);
   deviance_ratios.resize(k);
@@ -409,8 +415,12 @@ List cppSLOPE(T& x, mat& y, const List control)
           y_scale,
           intercept);
 
-  // standardize lambda
-  lambda /= n;
+  // rescale alpha depending on standardization settings
+  if (scale == "l2") {
+    alpha /= sqrt(n);
+  } else if (scale == "sd" || scale == "none") {
+    alpha /= n;
+  }
 
   return List::create(
     Named("betas")               = wrap(betas),
@@ -423,7 +433,7 @@ List cppSLOPE(T& x, mat& y, const List control)
     Named("violations")          = wrap(violation_list),
     Named("deviance_ratio")      = wrap(deviance_ratios),
     Named("null_deviance")       = wrap(null_deviance),
-    Named("sigma")               = wrap(sigma),
+    Named("alpha")               = wrap(alpha),
     Named("lambda")              = wrap(lambda)
   );
 }